Move UUID to utils
