+++ auth/sign-up/Environment.test.ts
@@ -2,7 +2,7 @@ import { EmailAddress, Password, USPhoneNumber } from ".."
 import { createSignUpEnvironment } from "./Environment"
 import { TiFAPI, createTiFAPIFetch } from "@api-client"
 import { rest } from "msw"
-import { uuidString } from "@lib/UUID"
+import { uuidString } from "@lib/utils/UUID"
 import { TestCognitoError } from "@auth/CognitoHelpers"
 import { mswServer } from "@test-helpers/msw"
 import { UserHandle } from "@content-parsing"
